is-bgp-safe: Add `IsBgpSafeChecker` trait + impl in `tor-relay-selection`
Trait `IsBgpSafeChecker` contains the logic to determine if a guard relay is considered BGP safe.

Impl `IsBgpSafeCheckerCsv` use two data sources, one to load the safe ASNs and one to create the ip2asn mapper. Method `IsBgpSafeChecker::is_bgp_safe()` takes a given IP of a relay as argument, maps the IP to the ASN, and returns true if that ASN is loaded.