feat: Make KeystoreEntry::new() public
Commit 6958b6c8 changed the way the `Keystore` trait works.
The `list` method must now return a `KeystoreEntry`. Before this change,
it was essentially impossible to implement keystores outside of
`tor-keymgr`. For 3rd party users of the Arti API that want to implement
a custom keystore, it became essentially impossible to do so.

To make this work again, this commit makes KeystoreEntry::new() public,
if the experimental-api feature is enabled.