ci: Allow cargo-audit check to fail.
This check can fail without our code changing due to new RUSTSEC issues
being added. As such, we shouldn't cause it to fail CI.

Our release instructions include checking for any warnings in CI, which
puts the burden on the person doing the release to ensure that this
passes at least every month, and running it in CI ensures that it is
generally visible to people.